fix: circular dependency, ssh key not required when editing
