+++ src/utils/secrets/secret-utils.ts
@@ -21,7 +21,6 @@ import {
   SourceSecretType,
   BuildTimeSecret,
 } from '../../types';
-import { createK8sSecretResource } from '../create-utils';
 
 export { SecretForComponentOption };
 
@@ -278,6 +277,29 @@ export const getAnnotationForSecret = (values: AddSecretFormValues): { [key: str
   return { [SecretLabels.REPO_ANNOTATION]: values.source.repo };
 };
 
+export const createK8sSecretResource = (
+  values: AddSecretFormValues,
+  secretResource: SecretKind,
+): SecretKind => {
+  const labels = {
+    secret: getLabelsForSecret(values),
+  };
+
+  const annotations = getAnnotationForSecret(values);
+  const k8sSecretResource = {
+    ...secretResource,
+    metadata: {
+      ...secretResource.metadata,
+      labels: {
+        ...labels?.secret,
+      },
+      annotations,
+    },
+  };
+
+  return k8sSecretResource;
+};
+
 export const statusFromConditions = (
   conditions: SecretCondition[],
 ): RemoteSecretStatusReason | string => {
@@ -329,7 +351,7 @@ export const createSecretResource = async (
     resource: secretResource,
   });
 
-const getUpdatedSecretResource = async (newK8sSecretResource: SecretKind) => {
+const updateSecretResource = async (newK8sSecretResource: SecretKind) => {
   return await K8sQueryPatchResource({
     model: SecretModel,
     queryOptions: {
@@ -374,7 +396,7 @@ export const editSecretResource = (
         : updatedSecret.source['ssh-privatekey'];
   }
 
-  return getUpdatedSecretResource(newK8sSecretResource);
+  return updateSecretResource(newK8sSecretResource);
 };
 
 export const getSecretBreadcrumbs = (namespace: string, operation: string) => {
